Improve particles

+++ Asteroids/Engine/ParticleSystem/Particle/Particle.h
@@ -18,11 +18,11 @@ namespace Engine {
         bool alive;
         int age;
         int maxAge;
-        double size;
+        float size;
         ParticleFunc colorFunc;
 
     public:
-        Particle(float x, float y, float dx, float dy, int age, int maxAge, double size, ParticleFunc colorFunc);
+        Particle(float x, float y, float dx, float dy, int age, int maxAge, float size, ParticleFunc colorFunc);
         bool IsAlive();
         void Update(float maxX, float maxY, bool wrap = false);
         void Draw(Shader& shader, float width, float height);
